refactor: carbon, info service 분리
