+++ src/main/java/smu/earthranger/controller/FollowController.java
@@ -2,7 +2,6 @@ package smu.earthranger.controller;
 
 
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
