+++ src/main/java/smu/earthranger/service/RankService.java
@@ -11,9 +11,7 @@ import smu.earthranger.repository.FollowRepository;
 import smu.earthranger.repository.MemberRepository;
 
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
